feat: myPage header 수정 및 api
