+++ src/apis/class.ts
@@ -31,42 +31,3 @@ export const getClassVideo = async (classId: number) => {
   const response = await client.get(`/danceclasses/${classId}/video`);
   return response.data;
 };
-
-//클래스 정보 업로드
-export const postClassInfo = async (input: any) => {
-  const response = await client.post('/mypage/class', {
-    date: input.date,
-    days: {
-      fri: input.friday,
-      mon: input.monday,
-      sat: input.saturday,
-      sun: input.sunday,
-      thu: input.thursday,
-      tue: input.tuesday,
-      wed: input.wednesday,
-    },
-    detail1: input.classContent,
-    detail2: input.classUser,
-    detail3: input.classIntro,
-    difficulty: input.classLevel,
-    endTime: input.endTime,
-    genres: [
-      {
-        genre: input.genre,
-      },
-    ],
-    hashTag1: input.hashTag1,
-    hashTag2: input.hashTag2,
-    hashTag3: input.hashTag3,
-    location: input.location,
-    maxPeople: input.maxPeople,
-    method: input.classWay,
-    reserveLink: input.classLink,
-    song: input.classSong,
-    startTime: input.startTime,
-    title: input.title,
-    tuition: input.classFee,
-    videoFile: input.url,
-  });
-  return response.data;
-};
