+++ src/apis/dancer.ts
@@ -9,3 +9,21 @@ export const getFilteredDancerList = async (typingValue: any) => {
   });
   return response.data;
 };
+
+//댄서경력 조회
+export const getDancerPortfolio = async (id: any) => {
+  const response = await client.get('/profile/${id}/portfolio');
+  return response.data;
+};
+
+//댄서 포트폴리오 영상
+export const getDancerPortfolioVideo = async (id: any) => {
+  const response = await client.get('/profile/${id}/portfolio/video');
+  return response.data;
+};
+
+//댄서 수업
+export const getDancerClass = async (id: any) => {
+  const response = await client.get('/profile/${id}/class');
+  return response.data;
+};
