+++ src/styles/components/common/MyPageHeader.module.css
@@ -1,21 +1,25 @@
 .container {
-  background: var(--color-gray-700-bg);
-  opacity: 30%;
+  background: transparent;
+
   width: 100%;
-  height: 55px;
+  height: 85px;
 
   position: fixed;
   z-index: 999;
 }
 
 .containerBtn {
-  background: transparent;
-  padding: 0 18px;
+  background: rgb(23, 23, 27, 0.3);
   width: 100%;
-  height: 85px;
+  height: 55px;
+  padding: 0 18px;
 
   display: flex;
   justify-content: space-between;
+  align-items: center;
+
+  top: 0;
+  left: 0;
 
   position: fixed;
   z-index: 1000;
@@ -25,15 +29,6 @@
   cursor: pointer; /* -> 웹에서만 필요*/
   display: flex;
   justify-content: center;
-  position: relative;
-}
-
-.backBtn {
-  margin-top: 11.5px;
-}
-
-.dotBtn {
-  margin-top: 18.5px;
 }
 
 .logoutBtn {
@@ -45,8 +40,8 @@
   border-radius: 5px;
   background: var(--color-gray-500-card);
   color: var(--color-gray-100);
-  z-index: 999;
+  z-index: 1000;
   position: absolute;
   right: 18px;
-  margin-top: 44px;
+  margin-top: 69.5px;
 }
